Update day of week examples
The Unicode LDML documentation used to call for the weekday abbreviation to sometimes be more than three characters (such as `Tuesday` being represented as `Tues`). The documentation was corrected in 2015 to show using a three character abbreviation (in this example, `Tue`). (This change was made in the official docs in https://github.com/unicode-org/cldr/commit/ca91e67020308739ba9aafe262dee7435ec08d16).

While I was here, I also added the missing "short" form which is two characters.